fix: signup errors, profile modes changes
