+++ src/components/Form/Form.css
@@ -183,6 +183,7 @@
 }
 
 .button-error {
+    text-align: center;
     margin-top: 49px;
     margin-bottom: 8px;
 }
